requirements.txt
# Core dependencies
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
PyYAML>=6.0

# Machine Learning
torch>=2.0.0
sentence-transformers>=2.2.0

# Vector database
faiss-cpu>=1.7.4  # o faiss-gpu si tienes GPU

# API clients
requests>=2.31.0

# Progress bars
tqdm>=4.65.0

# Data handling
pandas>=2.0.0

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Notebooks (opcional)
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
